customizable id shift method for domain replication. usefull for applications with additional ids or id based particle connectivity, like molecules in MD apps
